# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Donor {
  id: number;
  firstName: string;
  lastName: string;
  bloodType: string;
  city: string;
  phone: string;
  lastDonationDate?: string | null;
  isEligible: boolean;
  totalDonations: number;
  createdAt: string;
}

export interface CreateDonorBody {
  firstName: string;
  lastName: string;
  bloodType: string;
  city: string;
  phone: string;
  dateOfBirth: string;
  lastDonationDate?: string | null;
  weight: number;
  hasChronicDisease: boolean;
  email?: string;
}

export interface BloodRequest {
  id: number;
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospitalName: string;
  city: string;
  urgency: string;
  status: string;
  contactPhone: string;
  notes?: string | null;
  createdAt: string;
  fulfilledAt?: string | null;
}

export interface CreateBloodRequestBody {
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospitalName: string;
  city: string;
  urgency: string;
  contactPhone: string;
  notes?: string | null;
}

export interface Hospital {
  id: number;
  name: string;
  city: string;
  address: string;
  phone: string;
  availableBloodTypes: string[];
  isBloodBank: boolean;
  latitude: number;
  longitude: number;
  operatingHours: string;
}

export type DashboardStatsBloodTypeBreakdownItem = {
  bloodType: string;
  count: number;
};

export interface DashboardStats {
  totalDonors: number;
  totalDonations: number;
  activeRequests: number;
  partnerHospitals: number;
  livesSaved: number;
  donationsThisMonth: number;
  bloodTypeBreakdown: DashboardStatsBloodTypeBreakdownItem[];
}

export interface EmergencyRequest {
  id: number;
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospitalName: string;
  city: string;
  urgency: string;
  contactPhone: string;
  minutesAgo: number;
  status: string;
}

export type ListDonorsParams = {
  bloodType?: string;
  city?: string;
  limit?: number;
  offset?: number;
};

export type ListDonors200 = {
  donors: Donor[];
  total: number;
};

export type ListBloodRequestsParams = {
  urgency?: string;
  bloodType?: string;
  status?: string;
  city?: string;
};

export type ListBloodRequests200 = {
  requests: BloodRequest[];
  total: number;
};

export type ListHospitalsParams = {
  city?: string;
  bloodType?: string;
};

export type ListHospitals200 = {
  hospitals: Hospital[];
  total: number;
};

export type GetEmergencyFeedParams = {
  city?: string;
  limit?: number;
};

export type GetEmergencyFeed200 = {
  emergencies: EmergencyRequest[];
};