# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List registered donors
 */
export const ListDonorsQueryParams = zod.object({
  bloodType: zod.coerce.string().optional(),
  city: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
  offset: zod.coerce.number().optional(),
});

export const ListDonorsResponse = zod.object({
  donors: zod.array(
    zod.object({
      id: zod.number(),
      firstName: zod.string(),
      lastName: zod.string(),
      bloodType: zod.string(),
      city: zod.string(),
      phone: zod.string(),
      lastDonationDate: zod.string().nullish(),
      isEligible: zod.boolean(),
      totalDonations: zod.number(),
      createdAt: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Register a new donor
 */
export const RegisterDonorBody = zod.object({
  firstName: zod.string(),
  lastName: zod.string(),
  bloodType: zod.string(),
  city: zod.string(),
  phone: zod.string(),
  dateOfBirth: zod.string(),
  lastDonationDate: zod.string().nullish(),
  weight: zod.number(),
  hasChronicDisease: zod.boolean(),
  email: zod.string().optional(),
});

/**
 * @summary List blood requests
 */
export const ListBloodRequestsQueryParams = zod.object({
  urgency: zod.coerce.string().optional(),
  bloodType: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
  city: zod.coerce.string().optional(),
});

export const ListBloodRequestsResponse = zod.object({
  requests: zod.array(
    zod.object({
      id: zod.number(),
      patientName: zod.string(),
      bloodType: zod.string(),
      unitsNeeded: zod.number(),
      hospitalName: zod.string(),
      city: zod.string(),
      urgency: zod.string(),
      status: zod.string(),
      contactPhone: zod.string(),
      notes: zod.string().nullish(),
      createdAt: zod.string(),
      fulfilledAt: zod.string().nullish(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Submit a blood request
 */
export const CreateBloodRequestBody = zod.object({
  patientName: zod.string(),
  bloodType: zod.string(),
  unitsNeeded: zod.number(),
  hospitalName: zod.string(),
  city: zod.string(),
  urgency: zod.string(),
  contactPhone: zod.string(),
  notes: zod.string().nullish(),
});

/**
 * @summary List hospitals and blood banks
 */
export const ListHospitalsQueryParams = zod.object({
  city: zod.coerce.string().optional(),
  bloodType: zod.coerce.string().optional(),
});

export const ListHospitalsResponse = zod.object({
  hospitals: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      city: zod.string(),
      address: zod.string(),
      phone: zod.string(),
      availableBloodTypes: zod.array(zod.string()),
      isBloodBank: zod.boolean(),
      latitude: zod.number(),
      longitude: zod.number(),
      operatingHours: zod.string(),
    }),
  ),
  total: zod.number(),
});

/**
 * @summary Get platform statistics
 */
export const GetDashboardStatsResponse = zod.object({
  totalDonors: zod.number(),
  totalDonations: zod.number(),
  activeRequests: zod.number(),
  partnerHospitals: zod.number(),
  livesSaved: zod.number(),
  donationsThisMonth: zod.number(),
  bloodTypeBreakdown: zod.array(
    zod.object({
      bloodType: zod.string(),
      count: zod.number(),
    }),
  ),
});

/**
 * @summary Get live emergency blood requests
 */
export const GetEmergencyFeedQueryParams = zod.object({
  city: zod.coerce.string().optional(),
  limit: zod.coerce.number().optional(),
});

export const GetEmergencyFeedResponse = zod.object({
  emergencies: zod.array(
    zod.object({
      id: zod.number(),
      patientName: zod.string(),
      bloodType: zod.string(),
      unitsNeeded: zod.number(),
      hospitalName: zod.string(),
      city: zod.string(),
      urgency: zod.string(),
      contactPhone: zod.string(),
      minutesAgo: zod.number(),
      status: zod.string(),
    }),
  ),
});