# lib/api-zod/src/generated/types/bloodRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */

export interface BloodRequest {
  id: number;
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospitalName: string;
  city: string;
  urgency: string;
  status: string;
  contactPhone: string;
  notes?: string | null;
  createdAt: string;
  fulfilledAt?: string | null;
}