# lib/api-zod/src/generated/types/createBloodRequestBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBloodRequestBody {
  patientName: string;
  bloodType: string;
  unitsNeeded: number;
  hospitalName: string;
  city: string;
  urgency: string;
  contactPhone: string;
  notes?: string | null;
}