# lib/api-zod/src/generated/types/createDonorBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateDonorBody {
  firstName: string;
  lastName: string;
  bloodType: string;
  city: string;
  phone: string;
  dateOfBirth: string;
  lastDonationDate?: string | null;
  weight: number;
  hasChronicDisease: boolean;
  email?: string;
}