# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsBloodTypeBreakdownItem } from "./dashboardStatsBloodTypeBreakdownItem";

export interface DashboardStats {
  totalDonors: number;
  totalDonations: number;
  activeRequests: number;
  partnerHospitals: number;
  livesSaved: number;
  donationsThisMonth: number;
  bloodTypeBreakdown: DashboardStatsBloodTypeBreakdownItem[];
}