# lib/api-zod/src/generated/types/donor.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MBC - Main Blood Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Donor {
  id: number;
  firstName: string;
  lastName: string;
  bloodType: string;
  city: string;
  phone: string;
  lastDonationDate?: string | null;
  isEligible: boolean;
  totalDonations: number;
  createdAt: string;
}